Implementing the price list.  Also fixed the price list interface to use long for ID

+++ src/main/java/com/hiklas/mucking/around/PriceList.java
@@ -4,6 +4,6 @@ package com.hiklas.mucking.around;
  * Created by fiona on 07/12/2016.
  */
 public interface PriceList {
-  ItemPrice getItemPrice(int id);
+  ItemPrice getItemPrice(long id);
 }
